casting double to integer division.

+++ src/edu/psu/compbio/seqcode/projects/naomi/xoqualitycontrol/GCbiasProfiler.java
@@ -42,9 +42,9 @@ public class GCbiasProfiler {
 		Iterator<Region> chroms = new ChromosomeGenerator<Genome>().execute(genome);
 		int chromNum = genome.getChromList().size();
 		
-		GCvsFragmentRate = new double [l+1][2];
+		GCvsFragmentRate = new double[l+1][2];
 		for (int i =0; i<l+1; i++){
-			GCvsFragmentRate[i][0] = i/l;
+			GCvsFragmentRate[i][0] = (double) (i/l);
 			GCvsFragmentRate[i][1] = 0;
 		}
 		
@@ -70,7 +70,7 @@ public class GCbiasProfiler {
 			
 			for (ControlledExperiment rep: manager.getReplicates()){
 				List<StrandedBaseCount> currentCounts = sampleCountsMap.get(rep.getSignal());
-				double [][] strandedCounts = new double [currchromSize][2];
+				double[][] strandedCounts = new double[currchromSize][2];
 				for (StrandedBaseCount hits :currentCounts){
 					if (hits.getStrand()=='+')
 						strandedCounts[hits.getCoordinate()][0]+=hits.getCount();
@@ -81,7 +81,7 @@ public class GCbiasProfiler {
 				currentCounts=null;
 			}			
 			
-			double [][] gcSinglePositionModel = new double [l+1][3];
+			double[][] gcSinglePositionModel = new double[l+1][3];
 			
 			System.out.println("current chrom is "+currChrom.getChrom());
